requirements.txt
tensorflow==2.12.0
pandas==1.5.0
numpy==1.23.0
pymongo==4.0.0
Flask==2.2.0
apache-airflow==2.3.0
